changed int64 to uint64 in constants

+++ core/constants.go
@@ -5,10 +5,10 @@ import (
 	"math/big"
 )
 
-var EpochLength int64 = 1200
-var NumberOfStates int64 = 5
+var EpochLength uint64 = 1200
+var NumberOfStates uint64 = 5
 var ChainId = big.NewInt(132333505628089)
-var StateLength = uint64(EpochLength / NumberOfStates)
+var StateLength = EpochLength / NumberOfStates
 var MaxRetries uint = 8
 var NilHash = common.Hash{0x00}
 var BlockCompletionTimeout = 30
